requirements.txt
# Основное
fastapi
uvicorn[standard]
pydantic

# Телеграм-бот
aiogram~=3.20.0.post0
python-dotenv~=1.1.0

# База данных
psycopg2-binary
SQLAlchemy~=2.0.41
alembic
mysql-connector-python


# Redis
redis~=6.2.0
aioredis

# Админ-панель (Flask + Jinja2)
Flask
Jinja2
Flask-Login
Flask-WTF

# Платёжные системы
requests
stripe
paypalrestsdk

# Безопасность
PyJWT
passlib[bcrypt]
cryptography
httpx

# Вспомогательные
python-multipart
email-validator
loguru~=0.7.3
pytest~=8.4.0

motor~=3.7.1
pymongo~=4.13.2
testcontainers~=4.10.0
pytest-asyncio~=1.0.0
backend~=0.2.4.1